TestBasePersister and TestBuilder moved to spec Helpers
