+++ spec/helpers/test_builder/cloud_manager.rb
