+++ spec/helpers/test_builder/shared.rb
